+++ Specs/Core/PerspectiveFrustumSpec.js
@@ -181,13 +181,13 @@ describe('Core/PerspectiveFrustum', function() {
 
     it('get pixel dimensions throws without pixel ratio', function() {
         expect(function() {
-            return frustum.getPixelDimensions(1.0, 1.0, undefined, 1.0, new Cartesian2());
+            return frustum.getPixelDimensions(1.0, 1.0, 1.0, undefined, new Cartesian2());
         }).toThrowDeveloperError();
     });
 
     it('get pixel dimensions throws with pixel ratio less than or equal to zero', function() {
         expect(function() {
-            return frustum.getPixelDimensions(1.0, 1.0, 0.0, 1.0, new Cartesian2());
+            return frustum.getPixelDimensions(1.0, 1.0, 1.0, 0.0, new Cartesian2());
         }).toThrowDeveloperError();
     });
 
@@ -195,8 +195,8 @@ describe('Core/PerspectiveFrustum', function() {
         var dimensions = new Cartesian2(1.0, 1.0);
         var pixelRatio = 1.0;
         var distance = 1.0;
-        var pixelSize = frustum.getPixelDimensions(dimensions.x, dimensions.y, pixelRatio, distance, new Cartesian2());
-        var expected = frustum._offCenterFrustum.getPixelDimensions(dimensions.x, dimensions.y, pixelRatio, distance, new Cartesian2());
+        var pixelSize = frustum.getPixelDimensions(dimensions.x, dimensions.y, distance, pixelRatio, new Cartesian2());
+        var expected = frustum._offCenterFrustum.getPixelDimensions(dimensions.x, dimensions.y, distance, pixelRatio, new Cartesian2());
         expect(pixelSize.x).toEqual(expected.x);
         expect(pixelSize.y).toEqual(expected.y);
     });
@@ -205,8 +205,8 @@ describe('Core/PerspectiveFrustum', function() {
         var dimensions = new Cartesian2(1.0, 1.0);
         var pixelRatio = 2.0;
         var distance = 1.0;
-        var pixelSize = frustum.getPixelDimensions(dimensions.x, dimensions.y, pixelRatio, distance, new Cartesian2());
-        var expected = frustum._offCenterFrustum.getPixelDimensions(dimensions.x, dimensions.y, pixelRatio, distance, new Cartesian2());
+        var pixelSize = frustum.getPixelDimensions(dimensions.x, dimensions.y, distance, pixelRatio, new Cartesian2());
+        var expected = frustum._offCenterFrustum.getPixelDimensions(dimensions.x, dimensions.y, distance, pixelRatio, new Cartesian2());
         expect(pixelSize.x).toEqual(expected.x);
         expect(pixelSize.y).toEqual(expected.y);
     });