+++ Specs/Core/PerspectiveOffCenterFrustumSpec.js
@@ -169,13 +169,13 @@ describe('Core/PerspectiveOffCenterFrustum', function() {
 
     it('get pixel dimensions throws without pixel ratio', function() {
         expect(function() {
-            return frustum.getPixelDimensions(1.0, 1.0, undefined, 1.0, new Cartesian2());
+            return frustum.getPixelDimensions(1.0, 1.0, 1.0, undefined, new Cartesian2());
         }).toThrowDeveloperError();
     });
 
     it('get pixel dimensions throws with pixel ratio less than or equal to zero', function() {
         expect(function() {
-            return frustum.getPixelDimensions(1.0, 1.0, 0.0, 1.0, new Cartesian2());
+            return frustum.getPixelDimensions(1.0, 1.0, 1.0, 0.0, new Cartesian2());
         }).toThrowDeveloperError();
     });
 
@@ -186,7 +186,7 @@ describe('Core/PerspectiveOffCenterFrustum', function() {
     });
 
     it('get pixel dimensions with pixel ratio', function() {
-        var pixelSize = frustum.getPixelDimensions(1.0, 1.0, 2.0, 1.0, new Cartesian2());
+        var pixelSize = frustum.getPixelDimensions(1.0, 1.0, 1.0, 2.0, new Cartesian2());
         expect(pixelSize.x).toEqual(4.0);
         expect(pixelSize.y).toEqual(4.0);
     });